added template detail page fetch
